auction_service/requirements.txt
Flask==2.0.3
mysql-connector==2.2.9
python-dateutil==2.8.2
requests==2.27.1
